console error removed from homepage load, next bug in profile api call
