+++ client/src/api-helper.ts
@@ -142,9 +142,6 @@ export async function getProfile() {
   try {
     const response = await fetch('http://127.0.0.1:8000/profile', {
       method: 'GET',
-      headers: {
-        'Content-Type': 'application/json'
-      },
       credentials: 'include',
     });
     if (response.ok) {
